+++ foodb/CHANGELOG.md
@@ -1,4 +1,14 @@
+## 0.0.5
+* revert fix on change stream timeout
+
+## 0.0.4
+* fix change stream timeout
+
+## 0.0.3
+* Changed ChangeResult from Map<String, dynamic> to Doc<Map<String, dynamic>>
+
 ## 0.0.2
 * Fix .env
+
 ## 0.0.1
 * First release, implemented basic API for CRUD operation, change stream and replication.